requirements.txt
allure-python-commons==2.13.4
packaging==24.0
pluggy==1.4.0
pytest==8.1.1
selenium==4.19.0
urllib3==2.2.1
